+++ src/classes/__test__/Payload.unit.test.ts
@@ -1,3 +1,4 @@
+import { ValidationError } from '../../errors'
 import { Payload } from '../Payload'
 
 describe('Payload', () => {
@@ -40,11 +41,16 @@ describe('Payload', () => {
 				duration: 1000,
 				position: 60000,
 			}
-			jest.setSystemTime(new Date('1592156234000'))
+			const firstTime = new Date(1592156234000)
+			const secondTime = new Date(1592156235000)
+			jest.useFakeTimers()
+			jest.setSystemTime(firstTime)
 			const firstPayload = new Payload(parameters)
-			jest.setSystemTime(new Date('1592156235000'))
+			jest.setSystemTime(secondTime)
 			const secondPayload = new Payload(parameters)
-			expect(firstPayload.createdAt).not.toEqual(secondPayload.createdAt)
+			jest.useRealTimers()
+			expect(firstPayload.createdAt).toEqual(firstTime.toISOString())
+			expect(secondPayload.createdAt).toEqual(secondTime.toISOString())
 		})
 	})
 
@@ -66,10 +72,10 @@ describe('Payload', () => {
 
 	describe('deserialize', () => {
 		it('should error when deserializing an empty serialized object', () => {
-			expect(() => Payload.deserialize('{}')).toThrow()
+			expect(() => Payload.deserialize('{}')).toThrow(ValidationError)
 		})
 		it('should error when deserializeing a partially populated serialized payload', () => {
-			expect(() => Payload.deserialize('{"type":"WHOOPS"}')).toThrow()
+			expect(() => Payload.deserialize('{"type":"WHOOPS"}')).toThrow(ValidationError)
 		})
 		it('should properly deserialize a serialized payload', () => {
 			const parameters = {
@@ -95,7 +101,7 @@ describe('Payload', () => {
 				position: 60000,
 			}
 			const serializedValues = JSON.stringify(parameters)
-			expect(() => Payload.deserialize(serializedValues)).toThrow()
+			expect(() => Payload.deserialize(serializedValues)).toThrow(ValidationError)
 		})
 	})
 })