+++ src/tools/ajv.ts
@@ -1,0 +1,5 @@
+import Ajv from 'ajv'
+import AjvKeywords from 'ajv-keywords'
+
+export const ajv = new Ajv()
+AjvKeywords(ajv, 'instanceof')