+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The `Payload` class is now part of the countertop.  This will replace the `@tvkitchen/base-classes` version of Payload.
 - TypeScript definitions related to `Payload`, `PayloadType`, and `PayloadParameters` are also exported.
+- New `ValidationError` error.
 
 ## [0.4.1] - 2022-06-08
 ### Changed